+++ WideGame/Assets/Editor/MapPreviewEditor.cs
@@ -2,11 +2,11 @@
 using System.Collections;
 using UnityEditor;
 
-[CustomEditor (typeof (MapGenerator))]
-public class MapGeneratorEditor : Editor {
+[CustomEditor (typeof (MapPreview))]
+public class MapPreviewEditor : Editor {
 
 	public override void OnInspectorGUI() {
-		MapGenerator mapGen = (MapGenerator)target;
+		MapPreview mapGen = (MapPreview)target;
 
 		if (DrawDefaultInspector ()) {
 			if (mapGen.autoUpdate) {
