+++ WideGame/Assets/Editor/MapPreviewEditor.cs.meta
