--- WideGame/Assets/Scripts/MapDisplay.cs
@@ -1,21 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class MapDisplay : MonoBehaviour {
-
-	public Renderer textureRender;
-	public MeshFilter meshFilter;
-	public MeshRenderer meshRenderer;
-
-	public void DrawTexture(Texture2D texture) {
-		textureRender.sharedMaterial.mainTexture = texture;
-		textureRender.transform.localScale = new Vector3 (texture.width, 1, texture.height);
-	}
-
-	public void DrawMesh(MeshData meshData) {
-		meshFilter.sharedMesh = meshData.CreateMesh ();
-
-		meshFilter.transform.localScale = Vector3.one * FindObjectOfType<MapGenerator> ().meshSettings.meshScale;
-	}
-
-}
