--- WideGame/Assets/Scripts/MapGenerator.cs
@@ -1,138 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-using System;
-using System.Threading;
-using System.Collections.Generic;
-
-public class MapGenerator : MonoBehaviour {
-
-	public enum DrawMode {NoiseMap, Mesh, FalloffMap};
-	public DrawMode drawMode;
-
-	public MeshSettings meshSettings;
-	public HeightMapSettings heightMapSettings;
-	public TextureData textureData;
-
-	public Material terrainMaterial;
-
-
-
-	[Range(0,MeshSettings.numSupportedLODs-1)]
-	public int editorPreviewLOD;
-	public bool autoUpdate;
-
-	float[,] falloffMap;
-
-	Queue<MapThreadInfo<HeightMap>> heightMapThreadInfoQueue = new Queue<MapThreadInfo<HeightMap>>();
-	Queue<MapThreadInfo<MeshData>> meshDataThreadInfoQueue = new Queue<MapThreadInfo<MeshData>>();
-
-
-	void Start() {
-		textureData.ApplyToMaterial (terrainMaterial);
-		textureData.UpdateMeshHeights (terrainMaterial, heightMapSettings.minHeight, heightMapSettings.maxHeight);
-	}
-
-	void OnValuesUpdated() {
-		if (!Application.isPlaying) {
-			DrawMapInEditor ();
-		}
-	}
-
-	void OnTextureValuesUpdated() {
-		textureData.ApplyToMaterial (terrainMaterial);
-	}
-	 
-
-	public void DrawMapInEditor() {
-		textureData.UpdateMeshHeights (terrainMaterial, heightMapSettings.minHeight, heightMapSettings.maxHeight);
-		HeightMap heightMap = HeightMapGenerator.GenerateHeightMap (meshSettings.numVertsPerLine, meshSettings.numVertsPerLine, heightMapSettings, Vector2.zero);
-
-		MapDisplay display = FindObjectOfType<MapDisplay> ();
-		if (drawMode == DrawMode.NoiseMap) {
-			display.DrawTexture (TextureGenerator.TextureFromHeightMap (heightMap.values));
-		} else if (drawMode == DrawMode.Mesh) {
-			display.DrawMesh (MeshGenerator.GenerateTerrainMesh (heightMap.values,meshSettings, editorPreviewLOD));
-		} else if (drawMode == DrawMode.FalloffMap) {
-			display.DrawTexture(TextureGenerator.TextureFromHeightMap(FalloffGenerator.GenerateFalloffMap(meshSettings.numVertsPerLine)));
-		}
-	}
-
-	public void RequestHeightMap(Vector2 centre, Action<HeightMap> callback) {
-		ThreadStart threadStart = delegate {
-			HeightMapThread (centre, callback);
-		};
-
-		new Thread (threadStart).Start ();
-	}
-
-	void HeightMapThread(Vector2 centre, Action<HeightMap> callback) {
-		HeightMap heightMap = HeightMapGenerator.GenerateHeightMap (meshSettings.numVertsPerLine, meshSettings.numVertsPerLine, heightMapSettings, centre);
-		lock (heightMapThreadInfoQueue) {
-			heightMapThreadInfoQueue.Enqueue (new MapThreadInfo<HeightMap> (callback, heightMap));
-		}
-	}
-
-	public void RequestMeshData(HeightMap heightMap, int lod, Action<MeshData> callback) {
-		ThreadStart threadStart = delegate {
-			MeshDataThread (heightMap, lod, callback);
-		};
-
-		new Thread (threadStart).Start ();
-	}
-
-	void MeshDataThread(HeightMap heightMap, int lod, Action<MeshData> callback) {
-		MeshData meshData = MeshGenerator.GenerateTerrainMesh (heightMap.values,meshSettings, lod);
-		lock (meshDataThreadInfoQueue) {
-			meshDataThreadInfoQueue.Enqueue (new MapThreadInfo<MeshData> (callback, meshData));
-		}
-	}
-
-	void Update() {
-		if (heightMapThreadInfoQueue.Count > 0) {
-			for (int i = 0; i < heightMapThreadInfoQueue.Count; i++) {
-				MapThreadInfo<HeightMap> threadInfo = heightMapThreadInfoQueue.Dequeue ();
-				threadInfo.callback (threadInfo.parameter);
-			}
-		}
-
-		if (meshDataThreadInfoQueue.Count > 0) {
-			for (int i = 0; i < meshDataThreadInfoQueue.Count; i++) {
-				MapThreadInfo<MeshData> threadInfo = meshDataThreadInfoQueue.Dequeue ();
-				threadInfo.callback (threadInfo.parameter);
-			}
-		}
-	}
-
-
-	void OnValidate() {
-
-		if (meshSettings != null) {
-			meshSettings.OnValuesUpdated -= OnValuesUpdated;
-			meshSettings.OnValuesUpdated += OnValuesUpdated;
-		}
-		if (heightMapSettings != null) {
-			heightMapSettings.OnValuesUpdated -= OnValuesUpdated;
-			heightMapSettings.OnValuesUpdated += OnValuesUpdated;
-		}
-		if (textureData != null) {
-			textureData.OnValuesUpdated -= OnTextureValuesUpdated;
-			textureData.OnValuesUpdated += OnTextureValuesUpdated;
-		}
-
-	}
-
-	struct MapThreadInfo<T> {
-		public readonly Action<T> callback;
-		public readonly T parameter;
-
-		public MapThreadInfo (Action<T> callback, T parameter)
-		{
-			this.callback = callback;
-			this.parameter = parameter;
-		}
-
-	}
-
-}
-	
-
