+++ WideGame/Assets/Scripts/MapPreview.cs
@@ -1,0 +1,80 @@
+﻿using System.Collections;
+using UnityEngine;
+
+public class MapPreview : MonoBehaviour {
+
+	public Renderer textureRender;
+	public MeshFilter meshFilter;
+	public MeshRenderer meshRenderer;
+
+	public enum DrawMode { NoiseMap, Mesh, FalloffMap }
+	public DrawMode drawMode;
+
+	public MeshSettings meshSettings;
+	public HeightMapSettings heightMapSettings;
+	public TextureData textureData;
+
+	public Material terrainMaterial;
+
+	[Range (0, MeshSettings.numSupportedLODs - 1)]
+	public int editorPreviewLOD;
+	public bool autoUpdate;
+
+	void OnValuesUpdated () {
+		if (!Application.isPlaying) {
+			DrawMapInEditor ();
+		}
+	}
+
+	void OnTextureValuesUpdated () {
+		textureData.ApplyToMaterial (terrainMaterial);
+	}
+
+	public void DrawMapInEditor () {
+		textureData.ApplyToMaterial (terrainMaterial);
+
+		textureData.UpdateMeshHeights (terrainMaterial, heightMapSettings.minHeight, heightMapSettings.maxHeight);
+		HeightMap heightMap = HeightMapGenerator.GenerateHeightMap (meshSettings.numVertsPerLine, meshSettings.numVertsPerLine, heightMapSettings, Vector2.zero);
+
+		if (drawMode == DrawMode.NoiseMap) {
+			DrawTexture (TextureGenerator.TextureFromHeightMap (heightMap));
+		} else if (drawMode == DrawMode.Mesh) {
+			DrawMesh (MeshGenerator.GenerateTerrainMesh (heightMap.values, meshSettings, editorPreviewLOD));
+		} else if (drawMode == DrawMode.FalloffMap) {
+			DrawTexture (TextureGenerator.TextureFromHeightMap (new HeightMap (FalloffGenerator.GenerateFalloffMap (meshSettings.numVertsPerLine), 0, 1)));
+		}
+	}
+
+	public void DrawTexture (Texture2D texture) {
+		textureRender.sharedMaterial.mainTexture = texture;
+		textureRender.transform.localScale = new Vector3 (texture.width, 1, texture.height) /10f;
+
+		meshFilter.gameObject.SetActive (false);
+		textureRender.gameObject.SetActive (true);
+	}
+
+	public void DrawMesh (MeshData meshData) {
+		meshFilter.sharedMesh = meshData.CreateMesh ();
+
+		meshFilter.gameObject.SetActive (true);
+		textureRender.gameObject.SetActive (false);
+	}
+
+	void OnValidate () {
+
+		if (meshSettings != null) {
+			meshSettings.OnValuesUpdated -= OnValuesUpdated;
+			meshSettings.OnValuesUpdated += OnValuesUpdated;
+		}
+		if (heightMapSettings != null) {
+			heightMapSettings.OnValuesUpdated -= OnValuesUpdated;
+			heightMapSettings.OnValuesUpdated += OnValuesUpdated;
+		}
+		if (textureData != null) {
+			textureData.OnValuesUpdated -= OnTextureValuesUpdated;
+			textureData.OnValuesUpdated += OnTextureValuesUpdated;
+		}
+
+	}
+
+}
