+++ WideGame/Assets/Scripts/TerrainChunk.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7fa5957b74ed41841b3525ae99c0074d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
