+++ WideGame/Assets/Scripts/TextureGenerator.cs
@@ -13,14 +13,14 @@ public static class TextureGenerator {
 	}
 
 
-	public static Texture2D TextureFromHeightMap(float[,] heightMap) {
-		int width = heightMap.GetLength (0);
-		int height = heightMap.GetLength (1);
+	public static Texture2D TextureFromHeightMap(HeightMap heightMap) {
+		int width = heightMap.values.GetLength (0);
+		int height = heightMap.values.GetLength (1);
 
 		Color[] colourMap = new Color[width * height];
 		for (int y = 0; y < height; y++) {
 			for (int x = 0; x < width; x++) {
-				colourMap [y * width + x] = Color.Lerp (Color.black, Color.white, heightMap [x, y]);
+				colourMap [y * width + x] = Color.Lerp (Color.black, Color.white, Mathf.InverseLerp(heightMap.minValue, heightMap.maxValue, heightMap.values[x,y]));
 			}
 		}
