+++ WideGame/Assets/Scripts/ThreadedDataRequester.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 916b4f4143ed7456294aa19afa97bde8
-timeCreated: 1454754067
-licenseType: Pro
+guid: 16089ea08f3423943a730a1472cda464
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
